added relation to task | alembic migration for user_id

+++ src/tasks/controller.py
@@ -3,19 +3,25 @@ from tkinter import NO
 from src.tasks.dtos import TaskSchema
 from sqlalchemy.orm import Session
 from src.tasks.models import TaskModel
+from src.user.models import UserModel
 from fastapi import HTTPException
 
-def create_task(body: TaskSchema, db: Session):
+def create_task(body: TaskSchema, db: Session, user:UserModel):
   data = body.model_dump()
-  new_task = TaskModel(title=data["title"], description = data["description"], is_completed = data["is_completed"])
+  new_task = TaskModel(
+      title=data["title"],
+      description = data["description"],
+      is_completed = data["is_completed"],
+      user_id = user.id
+    )
   db.add(new_task)
   db.commit()
   db.refresh(new_task)
   return new_task
 
 
-def get_task(db:Session):
-  tasks = db.query(TaskModel).all()
+def get_task(db:Session, user: UserModel):
+  tasks = db.query(TaskModel).filter(TaskModel.user_id == user.id).all()
   return tasks
 
 def get_one_task(task_id: int, db: Session):
@@ -24,26 +30,30 @@ def get_one_task(task_id: int, db: Session):
     raise HTTPException(404, detail="Task id not found")
   return one_task
 
-def update_task(body: TaskSchema, task_id: int, db: Session):
-  one_task = db.query(TaskModel).get(task_id)
+def update_task(body: TaskSchema, task_id: int, db: Session, user:UserModel):
+  one_task:TaskModel = db.query(TaskModel).get(task_id)
   if not one_task:
     raise HTTPException(404, detail="Task id not found")
+  
+  if one_task.user_id != user.id:
+    raise HTTPException(401, detail="You are not allowed to update this task")
+
   body = body.model_dump()
   for key, value in body.items():
     setattr(one_task, key, value)
-  # one_task.title = body.title
-  # one_task.description = body.description
-  # one_task.is_completed = body.is_completed
 
   db.add(one_task)
   db.commit()
   db.refresh(one_task)
   return  one_task
 
-def delete_task(task_id:int, db:Session):
-  one_task = db.query(TaskModel).get(task_id)
+def delete_task(task_id:int, db:Session, user: UserModel):
+  one_task:TaskModel = db.query(TaskModel).get(task_id)
   if not one_task:
     raise HTTPException(404, detail="Task id not found")
+  if one_task.user_id != user.id:
+    raise HTTPException(401, detail="You are not allowed to delete this task")
+
   db.delete(one_task)
   db.commit()
   return None
